Use session in all pages
